исправил согласно замечаниям

+++ Lessons7/src/main/java/com/artpro/model/AirTransport.java
@@ -12,8 +12,8 @@ public class AirTransport extends Transport {
     //В классе Воздушный добавляются поля:
 //- Размах крыльев (м)
 //- Минимальная длина взлётно-посадочной полосы для взлёта
-    private Integer wingspan; //размах крыльев
-    private Integer minRunwayLength; //длина взлётно-посадочной
+    private double wingspan; //размах крыльев
+    private double minRunwayLength; //длина взлётно-посадочной
 
 }
